requirements.txt
django==5.2
djangorestframework==3.15.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
